Non-compiling in progress splitting of the test file
