+++ commons/ava_testsuite/rpc_workflow_test/rpc_workflow_test_.go
@@ -1,0 +1,102 @@
+package rpc_workflow_test
+
+import (
+	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_networks"
+	"github.com/kurtosis-tech/kurtosis/commons/networks"
+	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
+	"github.com/palantir/stacktrace"
+	"github.com/sirupsen/logrus"
+	"strconv"
+	"time"
+)
+
+type StakingNetworkRpcWorkflowTest struct {
+	imageName string
+}
+func (test StakingNetworkRpcWorkflowTest) Run(network networks.Network, context testsuite.TestContext) {
+	castedNetwork := network.(ava_networks.TestGeckoNetwork)
+	stakerClient, err := castedNetwork.GetGeckoClient(NODE_SERVICE_ID)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get staker client"))
+	}
+	delegatorClient, err := castedNetwork.GetGeckoClient(DELEGATOR_NODE_SERVICE_ID)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get delegator client"))
+	}
+	stakerNodeId, err := stakerClient.InfoApi().GetNodeId()
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get staker node ID."))
+	}
+	delegatorNodeId, err := delegatorClient.InfoApi().GetNodeId()
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get delegator node ID."))
+	}
+	highLevelStakerClient := ava_networks.NewHighLevelGeckoClient(
+		stakerClient,
+		STAKER_USERNAME,
+		STAKER_PASSWORD)
+	highLevelDelegatorClient := ava_networks.NewHighLevelGeckoClient(
+		delegatorClient,
+		DELEGATOR_USERNAME,
+		DELEGATOR_PASSWORD)
+	stakerXchainAddress, err := highLevelStakerClient.CreateAndSeedXChainAccountFromGenesis(SEED_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not seed XChain account from Genesis."))
+	}
+	stakerPchainAddress, err := highLevelStakerClient.TransferAvaXChainToPChain(SEED_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not transfer AVA from XChain to PChain account information"))
+	}
+	_, err = highLevelDelegatorClient.CreateAndSeedXChainAccountFromGenesis(SEED_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not seed XChain account from Genesis."))
+	}
+	delegatorPchainAddress, err := highLevelDelegatorClient.TransferAvaXChainToPChain(SEED_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not transfer AVA from XChain to PChain account information"))
+	}
+	// Adding stakers
+	err = highLevelStakerClient.AddValidatorOnSubnet(stakerNodeId, stakerPchainAddress, STAKE_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not add staker %s to default subnet.", stakerNodeId))
+	}
+	currentStakers, err := stakerClient.PChainApi().GetCurrentValidators(nil)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get current stakers."))
+	}
+	logrus.Debugf("Number of current stakers: %d", len(currentStakers))
+	actualNumStakers := len(currentStakers)
+	expectedNumStakers := 6
+	context.AssertTrue(actualNumStakers == expectedNumStakers, stacktrace.NewError("Actual number of stakers, %v, != expected number of stakers, %v", actualNumStakers, expectedNumStakers))
+	// Adding delegators
+	err = highLevelDelegatorClient.AddDelegatorOnSubnet(stakerNodeId, delegatorPchainAddress, DELEGATOR_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not add delegator %s to default subnet.", delegatorNodeId))
+	}
+	/*
+		Currently no way to verify rewards for stakers and delegators because rewards are
+		only paid out at the end of the staking period, and the staking period must last at least
+		24 hours. This is far too long to be able to test in a CI scenario.
+	*/
+	remainingStakerAva := SEED_AMOUNT - STAKE_AMOUNT
+	_, err = highLevelStakerClient.TransferAvaPChainToXChain(stakerPchainAddress, stakerXchainAddress, remainingStakerAva)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to transfer Ava from PChain to XChain."))
+	}
+	xchainAccountInfo, err := stakerClient.XChainApi().GetBalance(stakerXchainAddress, ava_networks.AVA_ASSET_ID)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to get account info for account %v.", stakerXchainAddress))
+	}
+	actualRemainingAva := xchainAccountInfo.Balance
+	expectedRemainingAva := strconv.FormatInt(remainingStakerAva, 10)
+	context.AssertTrue(actualRemainingAva == expectedRemainingAva, stacktrace.NewError("Actual remaining Ava, %v, != expected remaining Ava, %v", actualRemainingAva, expectedRemainingAva))
+}
+func (test StakingNetworkRpcWorkflowTest) GetNetworkLoader() (networks.NetworkLoader, error) {
+	return getStakingNetworkLoader(map[int]int{
+		NODE_SERVICE_ID:           NORMAL_NODE_CONFIG_ID,
+		DELEGATOR_NODE_SERVICE_ID: NORMAL_NODE_CONFIG_ID,
+	}, test.imageName)
+}
+func (test StakingNetworkRpcWorkflowTest) GetTimeout() time.Duration {
+	return 90 * time.Second
+}
